requirements.txt
torch
torchvision
tensorboard
chardet
Pillow
tqdm